backend/requirements.txt
fastapi>=0.109.1
uvicorn>=0.24.0
motor>=3.3.1
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings==2.1.0
pymongo>=4.6.3
python-jose[cryptography]>=3.4.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18
requests==2.32.0